+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -5,11 +5,13 @@ CHANGELOG
 -----
 
  * deprecated the dialog helper (use the question helper instead)
+ * added a Process helper
  * deprecated TableHelper in favor of Table
  * deprecated ProgressHelper in favor of ProgressBar
  * added a question helper
  * added a way to set the process name of a command
  * added a way to set a default command instead of `ListCommand`
+ * added a way to set the process title of a command
 
 2.4.0
 -----